replace postdiffcallback class into adapter
